fix(core): remove hash prefix from cdn path (#5509)
The hash prefix will cause cache invalidate during deployment
It's for debug purpose but I forgot to remove it
